changed float column on integer

+++ db/migrate/20240819180301_add_discount_to_cart.rb
@@ -1,5 +1,5 @@
 class AddDiscountToCart < ActiveRecord::Migration[7.0]
   def change
-    add_column :carts, :discount, :float, default: 0.0
+    add_column :carts, :discount, :integer, default: 0
   end
 end
